application/requirements.txt
Flask==2.3.2
Flask==2.3.2
Flask-Login==0.6.2
Flask-Migrate==4.0.4
Flask-SQLAlchemy==3.0.5
Flask-WTF==1.1.1
SQLAlchemy==2.0.17
WTForms==3.0.1
SQLAlchemy==2.0.17
python-dotenv==1.0.0
pytest==7.4.0
Werkzeug==2.3.6
email-validator==2.0.0.post2
pandas==1.5.3
numpy==1.24.3
scikit-learn==1.2.2
joblib==1.2.0